% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimabs.R
\name{estimabs}
\alias{estimabs}
\title{Estimativa de maxima verossimilhanca dos parametros da distribuicao 
Birnbaum-Saunders (BS) para dados censurados do tipo II.}
\usage{
estimabs(x, delta = rep(1, times = length(x)))
}
\arguments{
\item{x}{Um vetor numerico.}

\item{delta}{Um vetor numerico que e 1 se a observacao e nao censurada
e 0 se for censurada.}
}
\value{
A estimativa dos parametros de forma \eqn{\alpha} e o parametro de escala
\eqn{\beta}.
}
\description{
A funcao calcula a estimativa de maxima verossimilhanca dos parametros
da distribuicao Birnbaum-Saunders.
}
\details{
Se o vetor \code{delta} nao for especificado, este assume por \emph{default}
um vetor apenas de 1 com comprimento igual ao do vetor \code{x}.
}
\references{
BIRNBAUM, Zygmund William; SAUNDERS, Sam C. 
A new family of life distributions. Journal of applied probability, 
v. 6, n. 2, p. 319-327, 1969.

BIRNBAUM, Zygmunt W.; SAUNDERS, Sam C.
 Estimation for a family of life distributions with 
 applications to fatigue. Journal of Applied Probability, v. 6, n. 2, p. 328-347, 1969.
}
\seealso{
\code{\link{maxLik}}.
}
\author{
Alan da Silva
}
