% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimaexp.R
\name{estimaexp}
\alias{estimaexp}
\title{Estimativa de maxima verossimilhanca dos parametros da distribuicao 
Exponencial (Exp) para dados censurados do tipo II.}
\usage{
estimaexp(x, delta = rep(1, times = length(x)))
}
\arguments{
\item{x}{Um vetor numerico.}

\item{delta}{Um vetor numerico que e 1 se a observacao e nao censurada
e 0 se for censurada.}
}
\value{
A estimativa do parametro de taxa \eqn{\alpha}.
}
\description{
A funcao calcula a estimativa de maxima verossimilhanca dos parametros
da distribuicao Exponencial.
}
\details{
Se o vetor \code{delta} nao for especificado, este assume por \emph{default}
um vetor apenas de 1 com comprimento igual ao do vetor \code{x}.
}
\seealso{
\code{\link{maxLik}}.
}
\author{
Alan da Silva
}
