% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVE.R
\name{HVEmn}
\alias{HVEmn}
\title{Estimativa da entropia populacional pelo estimador de Van Es(1992)}
\usage{
HVEmn(x)
}
\arguments{
\item{x}{Um vetor numerico.}
}
\value{
A estimativa da entropia com base no vetor \code{x}. 
\deqn{HVE_{mn}=1/(n-m) \sum _{i=1}^{n-m}log((n+1/m)[x_{i+m} - x_{i-m}]) + \sum _{k=m}^{n}(1/k) + log(m) - log(n+1)}
}
\description{
A funcao calcula a estimativa da entropia populacional
    com base na amostra.
}
\details{
O vetor \code{x} deve ter comprimento maior ou igual que 3.
}
\references{
VAN ES, Bert. Estimating functionals related to a density by a class of statistics based on spacings. 
Scandinavian Journal of Statistics, p. 61-72, 1992.
}
\author{
Alan da Silva
}
