% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HV.R
\name{HVmn}
\alias{HVmn}
\title{Estimativa da entropia populacional pelo estimador de Vasicek(1976)}
\usage{
HVmn(x)
}
\arguments{
\item{x}{Um vetor numerico.}
}
\value{
A estimativa da entropia com base no vetor \code{x}. \deqn{HV_{mn}=\sum _{i=1}^{n}log (n/2m)[x_{i+m} - x_{i-m}]}
}
\description{
A funcao calcula a estimativa da entropia populacional
    com base na amostra.
}
\details{
O vetor \code{x} deve ter comprimento maior ou igual que 3.
}
\references{
VASICEK, Oldrich. A test for normality based on sample entropy. Journal of the Royal
Statistical Society. Series B (Methodological), p. 54-59, 1976
}
\author{
Alan da Silva
}
