% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testvasicek.R
\name{testvasicek}
\alias{testvasicek}
\title{Teste de normalidade de Vasicek(1976).}
\usage{
testvasicek(x, nivel = 0.05, nrep = 10000)
}
\arguments{
\item{x}{e o vetor contendo a amostra.}

\item{nivel}{nivel de significancia nominal para testar a hipotese de normalidade.}

\item{nrep}{numero de repeticoes usadas na simulacao de Monte Carlo para a 
geracao da distribuicao amostral.}
}
\value{
A  estatistica de Vasicek, o valor critico ao nivel de
significancia estabelecido na funcao, bem como o p-valor obtido.
}
\description{
A funcao calcula a estatistica de teste, o valor critico e o p-valor
da estatistica de teste de Vasicek.
}
\details{
Se  \code{nrep} nao for especificado, este assume por \emph{default} 10.000.
De maneira analoga, se \code{nivel} nao for especificado, entao nivel=0,05.

A estatistica de Vasicek(1976) testa a hipotese de normalidade dos dados.
}
\references{
VASICEK, Oldrich. A test for normality based on sample entropy. 
Journal of the Royal Statistical Society. Series B (Methodological), p. 54-59, 1976.
}
\author{
Alan da Silva
}
