% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vasicekcrit.R
\name{vasicekcrit}
\alias{vasicekcrit}
\title{Valor critico da estatistica de Vasicek(1976).}
\usage{
vasicekcrit(n = length(x), nrep = 10000, nivel = 0.05)
}
\arguments{
\item{n}{tamanho da amostra.}

\item{nrep}{numero de repeticoes usadas na simulacao de Monte Carlo para a 
geracao da distribuicao amostral.}

\item{nivel}{nivel de significancia nominal para testar a hipotese de normalidade
da amostra.}
}
\value{
A distribuicao amostral da estatistica de Vasicek e o valor critico ao nivel de
significancia estabelecido na funcao.
}
\description{
A funcao calcula a distribuicao amostral da estatistica de
Vasicek, bem como o seu valor critico.
}
\details{
Se  \code{nrep} nao for especificado, este assume por \emph{default} 10.000.
De maneira analoga, se \code{nivel} nao for especificado, entao nivel=0,05.

A estatistica de Vasicek(1976) testa a hipotese de normalidade dos dados.
}
\references{
VASICEK, Oldrich. A test for normality based on sample entropy. 
Journal of the Royal Statistical Society. Series B (Methodological), p. 54-59, 1976.
}
\author{
Alan da Silva
}
